[.NET] Fix configuration drift between C# defaults and native config
Three discrepancies found by comparing C# hardcoded defaults against
zerobus_get_default_config() / zerobus_arrow_get_default_config():

1. CallbackMaxWaitTimeMs was null (wait forever), native default is 5000ms
2. MaxInflightBatches was 10000, native default is 1000
3. Arrow StreamPausedMaxWaitTimeMs was ulong (0), native expects int64_t
   with -1 meaning 'wait full server duration'

Also adds Defaults_MatchNativeConfig and ArrowDefaults_MatchNativeConfig
smoke tests to detect future drift.

+++ dotnet/src/Databricks.Zerobus/ZerobusSdk.cs
@@ -516,7 +516,7 @@ internal static class OptionsExtensions
             FlushTimeoutMs = (ulong)opts.FlushTimeoutMs,
             ConnectionTimeoutMs = (ulong)opts.ConnectionTimeoutMs,
             IpcCompression = (int)opts.IpcCompression,
-            StreamPausedMaxWaitTimeMs = (ulong)opts.StreamPausedMaxWaitTimeMs,
+            StreamPausedMaxWaitTimeMs = opts.StreamPausedMaxWaitTimeMs,
         };
     }
 }
